Revert "refactor: 사용자 로직 개선"
This reverts commit 7c04635689c48bb4f2a1cb4a104e38eef71e083f.
