+++ webtoon/src/main/kotlin/com/rlc/webtoon/dto/response/UserResponseDto.kt
@@ -4,7 +4,6 @@ import com.rlc.webtoon.entity.User
 import java.time.Instant
 
 data class UserResponseDto constructor(
-    val uuid: String,
     val id: String,
     val email: String,
     val phoneNumber: String,
@@ -15,8 +14,7 @@ data class UserResponseDto constructor(
     companion object {
         fun of(user: User): UserResponseDto {
             return UserResponseDto(
-                uuid = user.uuid!!,
-                id = user.loginId,
+                id = user.id,
                 email = user.email,
                 phoneNumber = user.phoneNumber,
                 birth = user.birth,
