+++ webtoon/src/main/kotlin/com/rlc/webtoon/entity/User.kt
@@ -1,7 +1,6 @@
 package com.rlc.webtoon.entity
 
 import com.rlc.webtoon.dto.request.UserRequestDto
-import com.rlc.webtoon.entity.common.BaseTimeEntity
 import jakarta.persistence.Column
 import jakarta.persistence.Entity
 import jakarta.persistence.Id
@@ -13,7 +12,7 @@ import java.util.UUID
 @Table(name = "tb_user")
 class User constructor(
     @Column(nullable = false)
-    val loginId: String,
+    val id: String,
     @Column(nullable = false)
     var password: String,
     @Column(nullable = false)
@@ -24,15 +23,13 @@ class User constructor(
     val birth: String,
 
     @Id
-    var uuid: String? = null,
-
-    var refreshToken: String? = null
+    private var uuid: String? = null
 ): BaseTimeEntity() {
 
     companion object {
         fun of(userRequestDto: UserRequestDto): User {
             return User(
-                loginId = userRequestDto.id,
+                id = userRequestDto.id,
                 password = userRequestDto.password,
                 email = userRequestDto.email,
                 phoneNumber = userRequestDto.phoneNumber,
@@ -41,9 +38,4 @@ class User constructor(
             )
         }
     }
-
-    override fun toString(): String {
-        return "User(loginId='$loginId', email='$email', phoneNumber='$phoneNumber', birth='$birth', uuid=$uuid, refreshToken=$refreshToken)"
-    }
-
 }
